+++ app/core/nice/event/mission.py
@@ -68,10 +68,6 @@ def get_nice_mission_cond(
     )
 
     if cond.condType == CondType.MISSION_CONDITION_DETAIL and cond.targetIds:
-        if cond.targetIds[0] in details:
-            nice_mission_cond.detail = get_nice_mission_cond_detail(
-                details[cond.targetIds[0]]
-            )
         nice_mission_cond.details = [
             get_nice_mission_cond_detail(details[detail_id])
             for detail_id in cond.targetIds